kmemdump: Add Kinfo backend driver
Add Kinfo backend driver.
This backend driver will select only regions of interest for the firmware,
and it copy those into a shared memory area that is supplied via OF.
The firmware is only interested in addresses for some symbols.
The list format is kinfo-compatible, with devices like Google Pixel phone.

Signed-off-by: Eugen Hristev <eugen.hristev@linaro.org>